VNA updates
